Archive Discussions (#9869)
## Summary

Discussion templates and workflows to effectively disable discussions in
this repo and direct people to an actively monitored community space.
This allows for faster support and more closely tracked feedback.

## Manual Verification

Not an SDK change. Just github yml files for discussion templates and
workflows.

## Checklist

<!--
Do not edit this list. Leave items unchecked that don't apply. If you
need to track subtasks, create a new "## Tasks" section

Please check if the PR fulfills the following requirements:
-->

- [ ] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [ ] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] I have reviewed this pull request (self-review)