fix(openai): allow 'revised_prompt' to be returned as null (#9891)
## Background

AI-SDK's internal Zod schema expects `revised_prompt` to always be a
non-null string. However, this is not always the case because
[revised_prompt is only returned when using the model `dall-e-3`
](https://platform.openai.com/docs/api-reference/images/object#images-object-data-revised_prompt)

As a result, the SDK throws a validation error.

#9441 

## Summary

set `revised_prompt` to have the nullish property

## Manual Verification

didn't need an e2e example. test has been updated to show where the
error was occurring

## Checklist

- [x] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [x] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)

## Related Issues

Fixes #9441