fix(test-server): make `msw` a production dependency (#9999)
## Background

We introduced a new package for the test server in #8588. The test
server always depended on `msw`, but we didn't add it as production
dependency because before we created a dedicated package for it, it was
part of provider-utils. Making `msw` a production dependency would have
meant that installing `ai` would install `msw` without ever needing it.

## Summary

`msw` will now automatically installed when installing
`@ai-sdk/test-server`. It's already fixed it in the `main` branch via
#9988

## Manual Verification

I'll try after the release but I'm sure enough this works as designed.

## Related Issues

#8588, #9988