fix(openai): treat unknown models as reasoning (#9976)
## Background

Most new models coming from OpenAI are reasoning now

## Summary

If a model is not recognized, it is now treated as a reasoning model.
I also refactored the code so that responses and chat models are sharing
the same logic.

## Manual Verification

tested with `openai` examples

## Future Work

Look into adding a flag to manually set if a model is treated as
reasoning.