feat(provider/google & provider/anthropic): support for custom provider name for anthropic and google (#10091)
## Background
With the baseURL of createAnthropic and createGoogleGenerativeAI, we can
use some providers that offer Anthropic or Gemini-compatible APIs.
Therefore, we hope to pass in a custom provider name to make it easier
to identify specific providers in logging and error tracking.

## Summary

Added support for custom provider name in google and anthropic

## Manual Verification
Wrote eg with telemetery for both google and anthropic to test the
custom provider names

## Related Issues

Fixes #9911

Co-authored-by: Gregor Martynus <39992+gr2m@users.noreply.github.com>