fix(provider)!: remove generics from EmbeddingModelV3 (#10592)
## Background

All embedding models that are used with the AI SDK are text embedding
models. The generic abstraction in the provider specification has been
premature. It makes our code unnecessarily complex and does not work
well with custom models and provider registries. If additional embedding
models (e.g. for images) are needed in the future, it would be better to
add a dedicated model type for those.

## Summary

* Remove generics from `EmbeddingModel`
* Remove generics from `embed` and `embedMany`
* Rename `textEmbedding` / `textEmbeddingModel` on providers to
`embedding` / `embeddingModel`

## Manual Verification

- [x] run `examples/ai-core/src/embed-many/openai.ts`
- [x] run `examples/ai-core/src/embed/openai.ts`

## Tasks

- [x] update provider
- [x] change all downstream usages
- [x] fix tests
- [x] update docs
- [x] changeset

## Future Work

* automated code migration for changing `textEmbedding` to `embedding`
(provider method)
* automated code migration for changing `textEmbeddingModel` to
`embeddingModel` (provider method)
* automatic code migration for custom provider constructor param

## Related Issues

Closes #9819

---------

Co-authored-by: Gregor Martynus <39992+gr2m@users.noreply.github.com>