fix(anthropic): throw error in doStream when the first stream chunk is an error (#10653)
## Background

The Anthropic API sends a 200 responses for streaming requests when
overloaded, and includes error chunks in the stream. This breaks the AI
SDK retry mechanism which relies on HTTP status codes and errors. While
this cannot be solved in general, because the expected state is unclear
when such an error happens mid-stream, it can be solved when the first
chunk is the error chunk (which may be the majority of cases).

## Summary

* move `DelayedPromise` into provider utils
* throw error from Anthropic language model `doStream` if first chunk is
an error

## Manual Verification

Cannot verify overload error end to end (Anthropic API would need to be
overloaded).

- [x] Run `examples/ai-core/src/stream-text/anthropic.ts` to ensure
streaming still works

## Future Work

* remove special case in case Anthropic switches to sending the correct
http status codes

## Related Issues

Fixes #2453