chore(openai): enable strict json by default (#10752)
## Background

OpenAI strict JSON ensures that valid, parseable JSON that matches a
schema is produced. However, it limits the supported JSON schemas. In AI
SDK 5 we turned off the strict JSON mode because of that. Unfortunately
it led to increased production stability issues (some schema generation
requests fail in production). To have increased stability in production,
we revert to enabling strict mode by default.

## Summary

* Turn on strict JSON schema generation by default for OpenAI responses
and chat models
* Remove `structuredOutputs` option from OpenAI chat models

## Related Issues
#8868