chore(examples) : next.js update for CVE-2025-66478 (#10876)
## Background

https://nextjs.org/blog/CVE-2025-66478

Since some of the examples currently fall within the scope of the
critical RCE vulnerability in React Server Components (CVE-2025-66478),
so I’d like to update this template as well to ensure it can be used
safely. I’d appreciate your consideration of this update.


## Summary

There are no functional changes.
In the `examples` folder, the Next.js version is updated from `15.5.4`
to `15.5.7`.
In the root `package.json`, the Next.js version is updated from
`15.0.0-canary.23` to `15.0.5`.


## Manual Verification

I tried several samples in `examples/next-openai` and confirmed that
they work without any issues.