fix(ai-sdk/react): Fix subscribeToMessages callback dependency in useChat (#10908)
## Background

Follow-up to:
- #7387

That PR fixed stale subscriptions when the chat ID changes. However, it
was incomplete since it only handled the case where the new ID was
passed in the options. It's also possible for you to pass in a `chat`
instance directly which could have a new ID.

## Summary

Tweak the logic there to use the `chatRef.current.id` in the dependency
array, so no matter how the ID changes we'll refresh the subscription
callbacks.

## Manual Verification

Didn't do manually testing since the unit tests for this specific case
(plus the one I added) seemed sufficient.