feat(provider/vertex): add express mode (#10931)
## Background

This adds express mode support to Vertex which is a simplified version
of authentication compared to the default OAuth flow.

## Summary

Express mode uses a different base URL and requires the api key to be
added as a search parameter. Since we need to bypass the default OAuth
flow I also added checks to the

## Manual Verification

Haven't been able to do this yet due to my Google account being old and
messed up. Still posting this PR in the meantime.