fix(vertex/anthropic): passing beta header only for structured outputs (#11015)
## Background

Currently, the vertex anthropic provider only checked the model ID and
would pass the beta header for all models that allow structured outputs.

But vertex doesn't yet support passing these beta headers. So we would
get an [error like this](https://github.com/vercel/ai/issues/10981) even
for a simple generateText function

This PR also addresses a bug when strict mode is enabled and the header
is passed (shouldn't happen)

## Summary

- only pass the beta header when actually using native output format and
when using strict tools

## Manual Verification

the example `google-vertex-anthropic-sonnet-4-5.ts` was used to verify
the fix (was failing before)

## Checklist

- [x] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [x] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)

## Future Work

ideally, we handle this shortcoming in our implementation of vertex
provider
- if beta header is to be passed, we intercept it and then we could pass
a providerOption saying `structuredOutputMode: 'jsonTool'` (continued in
this PR after the comment
https://github.com/vercel/ai/pull/11015#issuecomment-3634356904)

## Related Issues

Fixes #10981