fix(ai): skip tool input validation when output-error has undefined input (#11393)
## Background

if tool call input is invalid, ai sdk avoids setting `input` field and
sets `rawInput` field instead, problem is, the validation logic still
checks for `input` and expects it to be defined, even though it is
intentionally not defined when the tool input is invalid. this renders
methods like `pipeAgentUIStreamToResponse` which uses
`createAgentUIStream` unusable when the AI incorrectly sends invalid
input because internally it uses `validateUIMessages` which will always
fail because `input` field isn't set (copied from #11390 )

## Summary

Only validate when input is available since there are two possible
states for output-error the error happened before inputSchemas was
validated or after. Right now code always validates as it was after

## Manual Verification

Ran the test, and I'm fixing an issue from a live application, also the
issue provides a minimal reproduction.

## Related Issues

Fixes #11392
Similar to #11390

Co-authored-by: Lars Grammel <lars.grammel@gmail.com>